+++ idlib/hashing/CRC32.h
@@ -15,7 +15,7 @@ void CRC32_UpdateChecksum(unsigned int &crcvalue, const void *data, int length);
 void CRC32_FinishChecksum(unsigned int &crcvalue);
 unsigned int CRC32_BlockChecksum(const void *data, int length);
 #ifdef _SPLASHDAMAGE
-void CRC32_UpdateChecksum( unsigned long &crcvalue, const byte data );
+void CRC32_UpdateChecksum( unsigned /* 64long */int &crcvalue, const byte data );
 #endif
 
 #endif /* !__CRC32_H__ */
