+++ splashdamage/idlib/hashing/Honeyman.cpp
@@ -9,7 +9,7 @@
 
 #ifdef CREATE_CRC_TABLE
 
-static unsigned long crctable[256];
+static unsigned /* 64long */int crctable[256];
 
 /*
 	Create the CRC table for the simplified version of the pathalias hashing function.
@@ -48,7 +48,7 @@ void make_crc_table( void ) {
 
 #else
 
-static unsigned long crctable[256] = {
+static unsigned /* 64long */int crctable[256] = {
 	0x00000000L, 0x48000000L, 0x24000000L, 0x6c000000L,
 	0x12000000L, 0x5a000000L, 0x36000000L, 0x7e000000L,
 	0x09000000L, 0x41000000L, 0x2d000000L, 0x65000000L,
@@ -85,16 +85,16 @@ static unsigned long crctable[256] = {
 
 #endif
 
-void Honeyman_InitChecksum( unsigned long &crcvalue ) {
+void Honeyman_InitChecksum( unsigned /* 64long */int &crcvalue ) {
 	crcvalue = HONEYMAN_INIT_VALUE;
 }
 
-void Honeyman_UpdateChecksum( unsigned long &crcvalue, const byte data ) {
+void Honeyman_UpdateChecksum( unsigned /* 64long */int &crcvalue, const byte data ) {
 	crcvalue = ( ( crcvalue >> 7 ) ^ crctable[ ( crcvalue ^ data ) & 0x7f ] );
 }
 
-void Honeyman_UpdateChecksum( unsigned long &crcvalue, const void *data, int length ) {
-	unsigned long crc;
+void Honeyman_UpdateChecksum( unsigned /* 64long */int &crcvalue, const void *data, int length ) {
+	unsigned /* 64long */int crc;
 	const unsigned char *buf = (const unsigned char *) data;
 
 	crc = crcvalue;
@@ -104,12 +104,12 @@ void Honeyman_UpdateChecksum( unsigned long &crcvalue, const void *data, int len
 	crcvalue = crc;
 }
 
-void Honeyman_FinishChecksum( unsigned long &crcvalue ) {
+void Honeyman_FinishChecksum( unsigned /* 64long */int &crcvalue ) {
 	crcvalue ^= HONEYMAN_XOR_VALUE;
 }
 
-unsigned long Honeyman_BlockChecksum( const void *data, int length ) {
-	unsigned long crc;
+unsigned /* 64long */int Honeyman_BlockChecksum( const void *data, int length ) {
+	unsigned /* 64long */int crc;
 
 	Honeyman_InitChecksum( crc );
 	Honeyman_UpdateChecksum( crc, data, length );
